# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User Management System API",
    "description": "User Management System API",
    "contact": {
      "email": "bugrayaktiyol@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
...
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get a list of all users",
        "operationId": "listUsers",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserOutput"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a new user",
        "operationId": "createUser",
        "requestBody": {
          "description": "User object that needs to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutput"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserInput": {
        "required": [
          "address",
          "age",
          "name",
          "tags"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "age": {
            "type": "integer",
            "example": 25
          },
          "address": {
            "type": "string",
            "example": "123 Main St"
          },
          "tags": {
            "type": "string",
            "example": "cool,loser"
          }
        }
      },
      "UserOutput": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "integer",
            "example": 1027
          },
          "Name": {
            "type": "string",
            "example": "Nadean Schwimmer"
          },
          "Age": {
            "type": "integer",
            "example": 20
          },
          "Address": {
            "type": "string",
            "example": "15726 Express Crossing"
          },
          "Tags": {
            "type": "string",
            "example": "music,loser"
          },
          "CreatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "UpdatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "DeletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      }
    }
  }
}